tgbot_telethon/requirements.txt
pyyaml==6.0
telethon==1.24.0